src/json_producer/requirements.txt
azure-identity
azure-storage-blob
python-dotenv
Faker